Restructure of some print helper functions.
